fix(media): réparer les pages publiques de partage projet (v1.14.2)
Le fix v1.14.1 réparait les routes API /api/media/download et
/api/media/gallery pour les projets, mais les pages SSR publiques
(/media/d/[token], /media/g/[token]) dupliquaient la même logique
dans leur propre page.tsx sans jamais avoir été alignées : un lien
partagé sur un projet continuait d'afficher 0 média à l'ouverture.

La logique de résolution (événement → photos, projet → fichiers
validés) est centralisée dans resolveDownloadData/resolveGalleryData
(module media) et réutilisée par les routes API ET les pages SSR,
pour éviter toute nouvelle divergence entre les deux.
