+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "koinonia",
-  "version": "1.14.1",
+  "version": "1.14.2",
   "private": true,
   "scripts": {
     "dev": "next dev --turbopack",
