+++ src/app/media/d/[token]/page.tsx
@@ -3,40 +3,16 @@
  * Accessible via un lien MEDIA sans authentification.
  */
 import { notFound } from "next/navigation";
-import { prisma } from "@/lib/prisma";
-import { validateMediaShareToken, getSignedThumbnailUrl } from "@/modules/media";
+import { validateMediaShareToken, resolveDownloadData } from "@/modules/media";
 import DownloadView from "./DownloadView";
 
 async function fetchDownloadData(token: string) {
   try {
     const shareToken = await validateMediaShareToken(token, ["MEDIA", "MEDIA_ALL"]);
-    const event = shareToken.mediaEvent;
-
-    if (!event) return { token: shareToken, event: null, photos: [] };
-
-    const allStatuses = shareToken.type === "MEDIA_ALL";
-
-    const photos = await prisma.mediaPhoto.findMany({
-      where: { mediaEventId: event.id, ...(!allStatuses && { status: "APPROVED" }) },
-      orderBy: { uploadedAt: "asc" },
-    });
-
-    const photosWithUrls = await Promise.all(
-      photos.map(async (p) => ({
-        id: p.id,
-        filename: p.filename,
-        size: p.size,
-        width: p.width,
-        height: p.height,
-        status: p.status,
-        thumbnailUrl: await getSignedThumbnailUrl(p.thumbnailKey),
-      }))
-    );
-
+    const data = await resolveDownloadData(shareToken);
     return {
-      token: { id: shareToken.id, type: shareToken.type, label: shareToken.label },
-      event: { id: event.id, name: event.name, date: event.date.toISOString(), photoCount: photosWithUrls.length },
-      photos: photosWithUrls,
+      ...data,
+      event: data.event ? { ...data.event, date: data.event.date.toISOString() } : null,
     };
   } catch {
     return null;
