+++ src/app/media/g/[token]/page.tsx
@@ -3,38 +3,16 @@
  * Accessible via un lien GALLERY sans authentification.
  */
 import { notFound } from "next/navigation";
-import { prisma } from "@/lib/prisma";
-import { validateMediaShareToken, getSignedThumbnailUrl } from "@/modules/media";
+import { validateMediaShareToken, resolveGalleryData } from "@/modules/media";
 import GalleryView from "./GalleryView";
 
 async function fetchGalleryData(token: string) {
   try {
     const shareToken = await validateMediaShareToken(token, "GALLERY");
-    const onlyApproved = (shareToken.config as { onlyApproved?: boolean } | null)?.onlyApproved ?? false;
-    const event = shareToken.mediaEvent;
-
-    if (!event) return { token: shareToken, event: null, photos: [] };
-
-    const photos = await prisma.mediaPhoto.findMany({
-      where: { mediaEventId: event.id, ...(onlyApproved ? { status: "APPROVED" } : {}) },
-      orderBy: { uploadedAt: "asc" },
-    });
-
-    const photosWithUrls = await Promise.all(
-      photos.map(async (p) => ({
-        id: p.id,
-        filename: p.filename,
-        status: p.status,
-        width: p.width,
-        height: p.height,
-        thumbnailUrl: await getSignedThumbnailUrl(p.thumbnailKey),
-      }))
-    );
-
+    const data = await resolveGalleryData(shareToken);
     return {
-      token: { id: shareToken.id, type: shareToken.type, label: shareToken.label, config: shareToken.config },
-      event: { id: event.id, name: event.name, date: event.date.toISOString(), photoCount: photosWithUrls.length },
-      photos: photosWithUrls,
+      ...data,
+      event: data.event ? { ...data.event, date: data.event.date.toISOString() } : null,
     };
   } catch {
     return null;
