+++ src/modules/media/index.ts
@@ -1,7 +1,7 @@
 import { defineModule } from "@/core/module-registry";
 
-export { createMediaShareToken, validateMediaShareToken, getTokenUrlPath, isTokenExpired, generateToken, collectionPhotoWhere } from "./services/tokens";
-export type { CollectionConfig } from "./services/tokens";
+export { createMediaShareToken, validateMediaShareToken, getTokenUrlPath, isTokenExpired, generateToken, collectionPhotoWhere, resolveDownloadData, resolveGalleryData } from "./services/tokens";
+export type { CollectionConfig, ResolvedMediaData, ResolvedGalleryData, ResolvedMediaEntry } from "./services/tokens";
 export { processImage, validatePhotoFile, getExtensionFromMimeType, ALLOWED_PHOTO_MIME_TYPES, MAX_PHOTO_SIZE } from "./services/image";
 export {
   uploadFile as uploadMediaFile,
